split the minio policy into individual class
