+++ common/__init__.py
@@ -2,4 +2,4 @@ from .geid import *
 from .logger import *
 from .vault import *
 from .project import *
-from .boto3 import *
+from .object_storage_adaptor import *
