+++ common/object_storage_adaptor/boto3_client.py
@@ -19,11 +19,18 @@ import xmltodict
 
 from botocore.client import Config
 
-
 _SIGNATURE_VERSTION = 's3v4'
 
 
-class Minio_Client:
+async def get_minio_client(minio_endpoint:str, token:str):
+
+    mc = Boto3_Client(minio_endpoint, token)
+    await mc.init_connection()
+
+    return mc
+
+
+class Boto3_Client:
     """
     Summary:
         The object client for minio operation. This class is based on
@@ -254,23 +261,3 @@ class Minio_Client:
             )
 
         return res
-
-
-async def get_minio_client(minio_endpoint:str, token:str) -> Minio_Client:
-    """
-    Summary:
-        The async function to get the minio client
-
-    Parameter:
-        - access_token(str): The token get from SSO
-        - duration(int): how long the temporary credential
-            will expire
-
-    return:
-        - dict
-    """
-
-    mc = Minio_Client(minio_endpoint, token)
-    await mc.init_connection()
-
-    return mc
